requirements.txt
streamlit>=1.31.0
supabase>=2.3.0
openai>=1.12.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0
python-docx>=1.1.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
numpy>=1.26.0
pandas>=2.1.0
plotly>=5.18.0
requests>=2.31.0
huggingface_hub>=0.20.0
